bgcheck: widen static-collision index types to 32-bit
A single scene was capped by three integer widths in static collision, and
the two tightest bit before the node ceiling anyone had noticed:

  node index  u16 -> u32   (SSNode.next, SSList.head, SS_NULL)
  poly id     s16 -> s32   (SSNode.polyId; polys past 32,768 went negative
                            and indexed polyList[] out of bounds at load)
  vertex id   13 -> 29 bits (flags_vIA/vIB pack the id under 3 flag bits)

All three ceilings now land at 65,535, which is what CollisionHeader's u16
numPolygons/numVertices can count. The full 6x6 F2P greybox at step 3
(32,856 polys, 16,697 vertices) was over two of the old three; it loads.

Nothing serializes SSNode/SSList, so those are a pure retype. The OTR
on-disk collision format is untouched and still packs at 13 bits -
CollisionHeaderFactory unpacks into the widened runtime fields at load, so
no vanilla scene data was migrated. vIC has no flags packed into it and is
widened by assignment alone.

Three sites read the packed flag bits by hand rather than through
COLPOLY_VIA_FLAG_TEST (SurfaceType_IsIgnoredByEntities /
_IsIgnoredByProjectiles / _IsConveyor) and would have silently kept testing
bits 13-15; they now name the flag and let the macro place it. COLPOLY_VTX
moves into z64bgcheck.h - it was duplicated in two custom-scene headers,
which is how the 13-bit assumption got missed twice. SOH::CollisionPoly is
a hand-kept copy of the engine struct that ResourceMgr_LoadColByName casts
between, so static_asserts in the factory now pin the two together.

Widening doubled SSNode and grew StaticLookup, and tblMax is whatever is
left of memSize after those fixed costs - so memSize doubles again to keep
every vanilla scene's node budget at least where it was. Measured after:
Forest Temple 12,284/50,609 nodes, Shadow Temple 11,543/49,322.

Refs sturdy-bassoon#22.

+++ soh/include/functions.h
@@ -577,7 +577,7 @@ s32 func_80037D98(PlayState* play, Actor* actor, s16 arg2, s32* arg3);
 s32 Actor_TrackPlayer(PlayState* play, Actor* actor, Vec3s* arg2, Vec3s* arg3, Vec3f arg4);
 
 // ? func_80038600(?);
-u16 DynaSSNodeList_GetNextNodeIdx(DynaSSNodeList*);
+u32 DynaSSNodeList_GetNextNodeIdx(DynaSSNodeList*);
 void func_80038A28(CollisionPoly* poly, f32 tx, f32 ty, f32 tz, MtxF* dest);
 f32 CollisionPoly_GetPointDistanceFromPlane(CollisionPoly* poly, Vec3f* point);
 CollisionHeader* BgCheck_GetCollisionHeader(CollisionContext* colCtx, s32 bgId);
@@ -646,7 +646,7 @@ s32 BgCheck_AnyLineTest3(CollisionContext* colCtx, Vec3f* posA, Vec3f* posB, Vec
 s32 BgCheck_SphVsFirstPoly(CollisionContext* colCtx, Vec3f* center, f32 radius);
 void SSNodeList_Initialize(SSNodeList*);
 void SSNodeList_Alloc(PlayState* play, SSNodeList* this, s32 tblMax, s32 numPolys);
-u16 SSNodeList_GetNextNodeIdx(SSNodeList* this);
+u32 SSNodeList_GetNextNodeIdx(SSNodeList* this);
 s32 DynaPoly_IsBgIdBgActor(s32 bgId);
 void DynaPoly_Init(PlayState* play, DynaCollisionContext* dyna);
 void DynaPoly_Alloc(PlayState* play, DynaCollisionContext* dyna);
